+++ src/matter/export.ts
@@ -1,4 +1,5 @@
 export * from '@matter/main';
+export { AttributeElement, ClusterElement, ClusterModel, CommandElement, EventElement, FieldElement } from '@matter/main/model';
 export * from '../matterbridgeDeviceTypes.js';
 export * from '../matterbridgeEndpoint.js';
 export * from '../matterbridgeBehaviors.js';